Refactor Common Models into a more consistent paradigm with more clearly defined semantics.
1. Contexts   - one way data from Loop to the Status & Watch extensions
2. Settings   - internal only models used to store Loop settings
3. Types      - largely internal enums
4. UserInfos  - dictionaries passed back from the watch to Loop
